pyproject.toml
[tool.poetry]
name = "vp"
version = "0.1.0"
description = ""
authors = ["Elias Yona <iconelias001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pymongo = "^4.3.3"
xlwt = "^1.3.0"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
